safari-600.7-branch/LayoutTests/sputnik/Conformance/12_Statement/12.11_The_switch_Statement/S12.11_A3_T5-expected.txt

CONSOLE MESSAGE: line 80: SyntaxError: Unexpected identifier 'result'. Expected '}' to end a body of a 'switch'.
S12.11_A3_T5

PASS Expected parsing failure

TEST COMPLETE
